Move Test Sim to Different File
